order_excel/requirements.txt
    pwdopenpyxl>=3.1
pandas>=2.0
python-barcode[images]>=0.15
Pillow>=10.0
requests>=2.31
python-telegram-bot>=20.0
aiohttp>=3.9